+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking
 
+- `get_page` and `get_section` no longer work with paths containing language code, eg "content/some.fr.md".
+
 ### Other
 
 - `get_taxonomy_url` `name` parameter is deprecated. Use `term` to be consistent with the other functions